+++ craft/app/services/SectionsService.php
@@ -6,8 +6,8 @@ namespace Craft;
  *
  * @author    Pixel & Tonic, Inc. <support@pixelandtonic.com>
  * @copyright Copyright (c) 2014, Pixel & Tonic, Inc.
- * @license   http://buildwithcraft.com/license Craft License Agreement
- * @see       http://buildwithcraft.com
+ * @license   http://craftcms.com/license Craft License Agreement
+ * @see       http://craftcms.com
  * @package   craft.app.services
  * @since     1.0
  */
@@ -15,11 +15,6 @@ class SectionsService extends BaseApplicationComponent
 {
 	// Properties
 	// =========================================================================
-
-	/**
-	 * @var
-	 */
-	public $typeLimits;
 
 	/**
 	 * @var
@@ -111,22 +106,10 @@ class SectionsService extends BaseApplicationComponent
 
 			$this->_sectionsById = array();
 
-			$typeCounts = array(
-				SectionType::Single => 0,
-				SectionType::Channel => 0,
-				SectionType::Structure => 0
-			);
-
 			foreach ($results as $result)
 			{
-				$type = $result['type'];
-
-				if (craft()->getEdition() >= Craft::Client || $typeCounts[$type] < $this->typeLimits[$type])
-				{
-					$section = new SectionModel($result);
-					$this->_sectionsById[$section->id] = $section;
-					$typeCounts[$type]++;
-				}
+				$section = new SectionModel($result);
+				$this->_sectionsById[$section->id] = $section;
 			}
 
 			$this->_fetchedAllSections = true;
@@ -338,11 +321,6 @@ class SectionsService extends BaseApplicationComponent
 		$sectionRecord->type             = $section->type;
 		$sectionRecord->enableVersioning = $section->enableVersioning;
 
-		if (($isNewSection || $section->type != $oldSection->type) && !$this->canHaveMore($section->type))
-		{
-			$section->addError('type', Craft::t('You can’t add any more {type} sections.', array('type' => Craft::t(ucfirst($section->type)))));
-		}
-
 		// Type-specific attributes
 		if ($section->type == SectionType::Single)
 		{
@@ -373,8 +351,16 @@ class SectionsService extends BaseApplicationComponent
 			$section->addError('localeErrors', Craft::t('At least one locale must be selected for the section.'));
 		}
 
+		$firstSectionLocale = null;
+
 		foreach ($sectionLocales as $localeId => $sectionLocale)
 		{
+			// Is this the first one?
+			if ($firstSectionLocale === null)
+			{
+				$firstSectionLocale = $sectionLocale;
+			}
+
 			if ($section->type == SectionType::Single)
 			{
 				$errorKey = 'urlFormat-'.$localeId;
@@ -610,14 +596,19 @@ class SectionsService extends BaseApplicationComponent
 					{
 						case SectionType::Single:
 						{
-							// In a nut, we want to make sure that there is one and only one Entry Type and Entry for this
-							// section. We also want to make sure the entry has rows in the i18n tables
-							// for each of the sections' locales.
-
+							// Make sure that there is one and only one Entry Type and Entry for this section.
 							$singleEntryId = null;
 
 							if (!$isNewSection)
 							{
+								// Re-save the entrytype name if the section name just changed
+								if (!$isNewSection && $oldSection->name != $section->name)
+								{
+									$entryType = $this->getEntryTypeById($entryTypeId);
+									$entryType->name = $section->name;
+									$this->saveEntryType($entryType);
+								}
+
 								// Make sure there's only one entry in this section
 								$entryIds = craft()->db->createCommand()
 									->select('id')
@@ -664,38 +655,12 @@ class SectionsService extends BaseApplicationComponent
 							if (!$singleEntryId)
 							{
 								// Create it, baby
-
-								craft()->db->createCommand()->insert('elements', array(
-									'type' => ElementType::Entry
-								));
-
-								$singleEntryId = craft()->db->getLastInsertID();
-
-								craft()->db->createCommand()->insert('entries', array(
-									'id'        => $singleEntryId,
-									'sectionId' => $section->id,
-									'typeId'    => $entryTypeId,
-									'postDate'  => DateTimeHelper::currentTimeForDb()
-								));
-							}
-
-							// Now make sure we've got all of the i18n rows in place.
-							foreach ($sectionLocales as $localeId => $sectionLocale)
-							{
-								craft()->db->createCommand()->insertOrUpdate('elements_i18n', array(
-									'elementId' => $singleEntryId,
-									'locale'    => $localeId,
-								), array(
-									'slug' => $section->handle,
-									'uri'  => $sectionLocale->urlFormat
-								));
-
-								craft()->db->createCommand()->insertOrUpdate('content', array(
-									'elementId' => $singleEntryId,
-									'locale'    => $localeId
-								), array(
-									'title' => $section->name
-								));
+								$singleEntry = new EntryModel();
+								$singleEntry->locale = $firstSectionLocale->locale;
+								$singleEntry->sectionId = $section->id;
+								$singleEntry->typeId = $entryTypeId;
+								$singleEntry->getContent()->title = $section->name;
+								craft()->entries->saveEntry($singleEntry);
 							}
 
 							break;
@@ -707,7 +672,7 @@ class SectionsService extends BaseApplicationComponent
 							{
 								// Add all of the entries to the structure
 								$criteria = craft()->elements->getCriteria(ElementType::Entry);
-								$criteria->locale = array_shift(array_keys($oldSectionLocales));
+								$criteria->locale = ArrayHelper::getFirstKey($oldSectionLocales);
 								$criteria->sectionId = $section->id;
 								$criteria->status = null;
 								$criteria->localeEnabled = null;
@@ -739,7 +704,7 @@ class SectionsService extends BaseApplicationComponent
 						$criteria = craft()->elements->getCriteria(ElementType::Entry);
 
 						// Get the most-primary locale that this section was already enabled in
-						$locales = array_intersect(craft()->i18n->getSiteLocaleIds(), array_keys($oldSectionLocales));
+						$locales = array_values(array_intersect(craft()->i18n->getSiteLocaleIds(), array_keys($oldSectionLocales)));
 
 						if ($locales)
 						{
@@ -811,48 +776,64 @@ class SectionsService extends BaseApplicationComponent
 			return false;
 		}
 
-		$transaction = craft()->db->getCurrentTransaction() === null ? craft()->db->beginTransaction() : null;
-		try
+		// Fire an 'onBeforeDeleteSection' event
+		$event = new Event($this, array(
+			'sectionId' => $sectionId
+		));
+
+		$this->onBeforeDeleteSection($event);
+
+		if ($event->performAction)
 		{
-			// Grab the entry ids so we can clean the elements table.
-			$entryIds = craft()->db->createCommand()
-				->select('id')
-				->from('entries')
-				->where(array('sectionId' => $sectionId))
-				->queryColumn();
+			$transaction = craft()->db->getCurrentTransaction() === null ? craft()->db->beginTransaction() : null;
 
-			craft()->elements->deleteElementById($entryIds);
-
-			// Delete the structure, if there is one
-			$structureId = craft()->db->createCommand()
-				->select('structureId')
-				->from('sections')
-				->where(array('id' => $sectionId))
-				->queryScalar();
-
-			if ($structureId)
+			try
 			{
-				craft()->structures->deleteStructureById($structureId);
+				// Grab the entry ids so we can clean the elements table.
+				$entryIds = craft()->db->createCommand()
+					->select('id')
+					->from('entries')
+					->where(array('sectionId' => $sectionId))
+					->queryColumn();
+
+				craft()->elements->deleteElementById($entryIds);
+
+				// Delete the structure, if there is one
+				$structureId = craft()->db->createCommand()
+					->select('structureId')
+					->from('sections')
+					->where(array('id' => $sectionId))
+					->queryScalar();
+
+				if ($structureId)
+				{
+					craft()->structures->deleteStructureById($structureId);
+				}
+
+				// Delete the section.
+				$affectedRows = craft()->db->createCommand()->delete('sections', array('id' => $sectionId));
+
+				if ($transaction !== null)
+				{
+					$transaction->commit();
+				}
+
+				// Fire an 'onDeleteSection' event
+				$this->onDeleteSection(new Event($this, array(
+					'sectionId' => $sectionId
+				)));
+
+				return (bool)$affectedRows;
 			}
-
-			// Delete the section.
-			$affectedRows = craft()->db->createCommand()->delete('sections', array('id' => $sectionId));
-
-			if ($transaction !== null)
+			catch (\Exception $e)
 			{
-				$transaction->commit();
-			}
+				if ($transaction !== null)
+				{
+					$transaction->rollback();
+				}
 
-			return (bool) $affectedRows;
-		}
-		catch (\Exception $e)
-		{
-			if ($transaction !== null)
-			{
-				$transaction->rollback();
+				throw $e;
 			}
-
-			throw $e;
 		}
 	}
 
@@ -1005,20 +986,26 @@ class SectionsService extends BaseApplicationComponent
 				// Is the event giving us the go-ahead?
 				if ($event->performAction)
 				{
-					if (!$isNewEntryType && $oldEntryType->fieldLayoutId)
+					// Is there a new field layout?
+					$fieldLayout = $entryType->getFieldLayout();
+
+					if (!$fieldLayout->id)
 					{
-						// Drop the old field layout
-						craft()->fields->deleteLayoutById($oldEntryType->fieldLayoutId);
+						// Delete the old one
+						if (!$isNewEntryType && $oldEntryType->fieldLayoutId)
+						{
+							craft()->fields->deleteLayoutById($oldEntryType->fieldLayoutId);
+						}
+
+						// Save the new one
+						craft()->fields->saveLayout($fieldLayout);
+
+						// Update the entry type record/model with the new layout ID
+						$entryType->fieldLayoutId = $fieldLayout->id;
+						$entryTypeRecord->fieldLayoutId = $fieldLayout->id;
 					}
 
-					// Save the new one
-					$fieldLayout = $entryType->getFieldLayout();
-					craft()->fields->saveLayout($fieldLayout);
-
-					// Update the entry type record/model with the new layout ID
-					$entryType->fieldLayoutId = $fieldLayout->id;
-					$entryTypeRecord->fieldLayoutId = $fieldLayout->id;
-
+					// Save the entry type
 					$entryTypeRecord->save(false);
 
 					// Now that we have an entry type ID, save it on the model
@@ -1216,37 +1203,6 @@ class SectionsService extends BaseApplicationComponent
 		return (bool) $count;
 	}
 
-	/**
-	 * Returns whether another section can be added of a given type.
-	 *
-	 * @param string $type
-	 *
-	 * @return bool
-	 */
-	public function canHaveMore($type)
-	{
-		if (craft()->getEdition() >= Craft::Client)
-		{
-			return true;
-		}
-		else
-		{
-			if (isset($this->typeLimits[$type]))
-			{
-				$count = craft()->db->createCommand()
-					->from('sections')
-					->where('type = :type', array(':type' => $type))
-					->count('id');
-
-				return $count < $this->typeLimits[$type];
-			}
-			else
-			{
-				return false;
-			}
-		}
-	}
-
 	// Events
 	// -------------------------------------------------------------------------
 
@@ -1296,6 +1252,30 @@ class SectionsService extends BaseApplicationComponent
 	public function onSaveSection(Event $event)
 	{
 		$this->raiseEvent('onSaveSection', $event);
+	}
+
+	/**
+	 * Fires an 'onBeforeDeleteSection' event.
+	 *
+	 * @param Event $event
+	 *
+	 * @return null
+	 */
+	public function onBeforeDeleteSection(Event $event)
+	{
+		$this->raiseEvent('onBeforeDeleteSection', $event);
+	}
+
+	/**
+	 * Fires an 'onDeleteSection' event.
+	 *
+	 * @param Event $event
+	 *
+	 * @return null
+	 */
+	public function onDeleteSection(Event $event)
+	{
+		$this->raiseEvent('onDeleteSection', $event);
 	}
 
 	// Private Methods
